fix compilation errors

+++ src/manifold/src/boolean3.cpp
@@ -92,9 +92,12 @@ SparseIndices Filter11(const Manifold::Impl &inP, const Manifold::Impl &inQ,
                        const SparseIndices &p1q2, const SparseIndices &p2q1) {
   ZoneScoped;
   SparseIndices p1q1(3 * p1q2.size() + 3 * p2q1.size());
-  for_each_n(autoPolicy(p1q2.size()), zip(countAt(0ul), countAt(0ul)),
-             p1q2.size(), CopyFaceEdges<false>({p1q2, p1q1, inQ.halfedge_}));
-  for_each_n(autoPolicy(p2q1.size()), zip(countAt(p1q2.size()), countAt(0ul)),
+  for_each_n(
+      autoPolicy(p1q2.size()),
+      zip(countAt(static_cast<size_t>(0)), countAt(static_cast<size_t>(0))),
+      p1q2.size(), CopyFaceEdges<false>({p1q2, p1q1, inQ.halfedge_}));
+  for_each_n(autoPolicy(p2q1.size()),
+             zip(countAt(p1q2.size()), countAt(static_cast<size_t>(0))),
              p2q1.size(), CopyFaceEdges<true>({p2q1, p1q1, inP.halfedge_}));
   p1q1.Unique();
   return p1q1;
@@ -261,7 +264,8 @@ std::tuple<Vec<int>, Vec<glm::vec4>> Shadow11(SparseIndices &p1q1,
   Vec<glm::vec4> xyzz11(p1q1.size());
 
   for_each_n(autoPolicy(p1q1.size()),
-             zip(countAt(0ul), xyzz11.begin(), s11.begin()), p1q1.size(),
+             zip(countAt(static_cast<size_t>(0)), xyzz11.begin(), s11.begin()),
+             p1q1.size(),
              Kernel11({inP.vertPos_, inQ.vertPos_, inP.halfedge_, inQ.halfedge_,
                        expandP, inP.vertNormal_, p1q1}));
 
@@ -352,7 +356,8 @@ std::tuple<Vec<int>, Vec<float>> Shadow02(const Manifold::Impl &inP,
 
   auto vertNormalP = forward ? inP.vertNormal_ : inQ.vertNormal_;
   for_each_n(autoPolicy(p0q2.size()),
-             zip(countAt(0ul), s02.begin(), z02.begin()), p0q2.size(),
+             zip(countAt(static_cast<size_t>(0)), s02.begin(), z02.begin()),
+             p0q2.size(),
              Kernel02({inP.vertPos_, inQ.halfedge_, inQ.vertPos_, expandP,
                        vertNormalP, p0q2, forward}));
 
@@ -456,7 +461,8 @@ std::tuple<Vec<int>, Vec<glm::vec3>> Intersect12(
   Vec<glm::vec3> v12(p1q2.size());
 
   for_each_n(
-      autoPolicy(p1q2.size()), zip(countAt(0ul), x12.begin(), v12.begin()),
+      autoPolicy(p1q2.size()),
+      zip(countAt(static_cast<size_t>(0)), x12.begin(), v12.begin()),
       p1q2.size(),
       Kernel12({p0q2.AsVec64(), s02, z02, p1q1.AsVec64(), s11, xyzz11,
                 inP.halfedge_, inQ.halfedge_, inP.vertPos_, forward, p1q2}));
